+++ app/models/user.rb
@@ -46,12 +46,6 @@ class User < ApplicationRecord
     likes.exists?(article_id: article.id)
   end
 
-  def display_name
-    # ぼっち演算子
-    # profileがnilじゃなかったらnicknameを使うとなる
-    profile&.nickname || self.email.split('@').first
-  end
-
   def follow!(user)
     user_id = get_user_id(user)
     following_relationships.create!(following_id: user_id)
@@ -66,17 +60,11 @@ class User < ApplicationRecord
   def has_followed?(user)
     following_relationships.exists?(following_id: user.id)
   end
+
   def prepare_profile
     profile || build_profile
   end
 
-  def avatar_image
-    if profile&.avater&.attached?
-      profile.avater
-    else
-      'default-avatar.png'
-    end
-  end
 
   private
   def get_user_id(user)
